Refactor of UpdateBuffer Level code
