+++ src/turtle_kv/tree/in_memory_node.cpp
@@ -23,12 +23,12 @@ namespace turtle_kv {
 namespace {
 
 using UpdateBuffer = InMemoryNode::UpdateBuffer;
-using Level = UpdateBuffer::Level;
-using EmptyLevel = UpdateBuffer::EmptyLevel;
-using MergedLevel = UpdateBuffer::MergedLevel;
-using SegmentedLevel = UpdateBuffer::SegmentedLevel;
-using HybridLevel = UpdateBuffer::HybridLevel;
-using Segment = UpdateBuffer::Segment;
+using Level = InMemoryNodeLevel;
+using EmptyLevel = InMemoryNodeEmptyLevel;
+using MergedLevel = InMemoryNodeMergedLevel;
+using SegmentedLevel = InMemoryNodeSegmentedLevel;
+using HybridLevel = InMemoryNodeHybridLevel;
+using Segment = InMemoryNodeSegment;
 
 using PackedUpdateBuffer = PackedNodePage::UpdateBuffer;
 using PackedLevel = PackedUpdateBuffer::SegmentedLevel;
@@ -1669,799 +1669,6 @@ StatusOr<llfs::PageId> InMemoryNode::finish_serialize(TreeSerializeContext& cont
 
 //==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
 //
-StatusOr<llfs::PinnedPage> Segment::load_leaf_page(llfs::PageLoader& page_loader,
-                                                   llfs::PinPageToJob pin_page_to_job,
-                                                   llfs::PageCacheOvercommit& overcommit) const
-{
-  return this->page_id_slot.load_through(page_loader,
-                                         llfs::PageLoadOptions{
-                                             LeafPageView::page_layout_id(),
-                                             pin_page_to_job,
-                                             llfs::OkIfNotFound{false},
-                                             llfs::LruPriority{kLeafLruPriority},
-                                             overcommit,
-                                         });
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-Level MergedLevel::merge(Level&& sibling_level, usize node_pivot_count) &&
-{
-  return batt::case_of(
-      sibling_level,
-      [&](EmptyLevel&) -> Level {
-        return *this;
-      },
-      [&](MergedLevel& right_merged_level) -> Level {
-        return this->concat(std::move(right_merged_level));
-      },
-      [&](auto& right_segmented_or_hybrid_level) -> Level {
-        HybridLevel new_hybrid_level;
-        new_hybrid_level.add_new_sub_level(std::move(*this));
-
-        new_hybrid_level.add_new_sub_level(std::move(right_segmented_or_hybrid_level),
-                                           node_pivot_count);
-
-        return new_hybrid_level;
-      });
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-StatusOr<usize> MergedLevel::start_serialize(const InMemoryNode& node,
-                                             TreeSerializeContext& context)
-{
-  batt::RunningTotal running_total =
-      compute_running_total(context.worker_pool(), this->result_set, DecayToItem<false>{});
-
-  SplitParts page_parts = split_parts(running_total,
-                                      MinPartSize{context.tree_options().flush_size() / 4},
-                                      MaxPartSize{context.tree_options().flush_size()},
-                                      MaxItemSize{context.tree_options().max_item_size()});
-
-  BATT_CHECK_EQ(running_total.back() - running_total.front(), this->result_set.get_packed_size());
-
-  auto filter_bits_per_key = context.tree_options().filter_bits_per_key();
-  const bool overcommit_triggered = context.overcommit().is_triggered();
-  llfs::PageSize filter_page_size = context.tree_options().filter_page_size();
-
-  for (const Interval<usize>& part_extents : page_parts) {
-    BATT_ASSIGN_OK_RESULT(
-        TreeSerializeContext::BuildPageJobId id,
-        context.async_build_page(
-            context.tree_options().leaf_size(),
-            packed_leaf_page_layout_id(),
-            llfs::LruPriority{kNewLeafLruPriority},
-            /*task_count=*/2,
-            [this,
-             &node,
-             part_extents,
-             filter_bits_per_key,
-             overcommit_triggered,
-             filter_page_size](
-                TreeSerializeContext::BuildPageArgs args) -> TreeSerializeContext::PinPageToJobFn {
-              //----- --- -- -  -  -   -
-              const auto all_items_in_level = this->result_set.get();
-              const auto items_in_this_page = batt::slice_range(all_items_in_level, part_extents);
-
-              if (args.task_i == 0) {
-                return build_leaf_page_in_job(node.tree_options.trie_index_reserve_size(),
-                                              args.page_buffer,
-                                              items_in_this_page);
-              }
-              BATT_CHECK_EQ(args.task_i, 1);
-
-              return build_filter_for_leaf_in_job(batt::make_copy(args.filter_page_write_state),
-                                                  args.page_cache,
-                                                  overcommit_triggered,
-                                                  filter_bits_per_key,
-                                                  filter_page_size,
-                                                  args.page_buffer.page_id(),
-                                                  items_in_this_page);
-            }));
-
-    this->segment_future_ids_.emplace_back(id);
-  }
-
-  return page_parts.size();
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-StatusOr<SegmentedLevel> MergedLevel::finish_serialize(const InMemoryNode& node,
-                                                       TreeSerializeContext& context)
-{
-  BATT_CHECK_EQ(node.tree_options.filter_bits_per_key(),
-                context.tree_options().filter_bits_per_key());
-  BATT_CHECK_EQ(node.tree_options.expected_items_per_leaf(),
-                context.tree_options().expected_items_per_leaf());
-
-  SegmentedLevel segmented_level;
-
-  const usize pivot_count = node.pivot_count();
-  const usize segment_count = this->segment_future_ids_.size();
-  segmented_level.segments.resize(segment_count);
-
-  for (usize segment_i = 0; segment_i < segment_count; ++segment_i) {
-    Segment& segment = segmented_level.segments[segment_i];
-
-    BATT_ASSIGN_OK_RESULT(llfs::PinnedPage pinned_leaf_page,
-                          context.get_build_page_result(this->segment_future_ids_[segment_i]));
-
-    segment.page_id_slot.page_id = pinned_leaf_page.page_id();
-    segment.active_pivots.clear();
-
-    const PackedLeafPage& leaf_page = PackedLeafPage::view_of(pinned_leaf_page);
-
-    for (usize pivot_i = 0; pivot_i < pivot_count; ++pivot_i) {
-      const Interval<KeyView> pivot_key_range = in_node(node).get_pivot_key_range(pivot_i);
-
-      const Interval<const PackedKeyValue*> pivot_range_in_leaf{
-          .lower_bound = leaf_page.lower_bound(pivot_key_range.lower_bound),
-          .upper_bound = leaf_page.lower_bound(pivot_key_range.upper_bound),
-      };
-
-      segment.set_pivot_active(pivot_i, !pivot_range_in_leaf.empty());
-    }
-
-    segment.check_invariants(__FILE__, __LINE__);
-  }
-
-  return {std::move(segmented_level)};
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-void InMemoryNode::UpdateBuffer::SegmentedLevel::drop_segment(usize i)
-{
-  this->segments.erase(this->segments.begin() + i);
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-void InMemoryNode::UpdateBuffer::SegmentedLevel::drop_pivot_range(
-    const Interval<i32>& pivot_i_range,
-    const Interval<KeyView>& pivot_key_range,
-    llfs::PageLoader& page_loader,
-    const TreeOptions& tree_options)
-{
-  for (Segment& segment : this->segments) {
-    BATT_CHECK_OK(in_segment(segment).drop_pivot_range(pivot_i_range,
-                                                       pivot_key_range,
-                                                       page_loader,
-                                                       tree_options));
-
-    if (pivot_i_range.lower_bound == 0) {
-      segment.pop_front_pivots(pivot_i_range.upper_bound);
-    }
-  }
-
-  this->segments.erase(std::remove_if(this->segments.begin(),
-                                      this->segments.end(),
-                                      [](const Segment& segment) {
-                                        return segment.is_inactive();
-                                      }),
-                       this->segments.end());
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-void InMemoryNode::UpdateBuffer::SegmentedLevel::drop_before_pivot(i32 pivot_i,
-                                                                   const KeyView& pivot_key,
-                                                                   llfs::PageLoader& page_loader,
-                                                                   const TreeOptions& tree_options)
-{
-  this->drop_pivot_range((Interval<i32>{0, pivot_i}),
-                         (Interval<KeyView>{global_min_key(), pivot_key}),
-                         page_loader,
-                         tree_options);
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-void InMemoryNode::UpdateBuffer::SegmentedLevel::drop_after_pivot(i32 pivot_i,
-                                                                  const KeyView& pivot_key,
-                                                                  llfs::PageLoader& page_loader,
-                                                                  const TreeOptions& tree_options)
-{
-  this->drop_pivot_range((Interval<i32>{pivot_i, InMemoryNode::kMaxTempPivots}),
-                         (Interval<KeyView>{pivot_key, global_max_key()}),
-                         page_loader,
-                         tree_options);
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-bool InMemoryNode::UpdateBuffer::SegmentedLevel::is_pivot_active(i32 pivot_i) const
-{
-  for (const Segment& segment : this->segments) {
-    if (segment.is_pivot_active(pivot_i)) {
-      return true;
-    }
-  }
-  return false;
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-void InMemoryNode::UpdateBuffer::SegmentedLevel::check_items_sorted(
-    const InMemoryNode& node,
-    llfs::PageLoader& page_loader) const
-{
-  SegmentedLevelScanner<const InMemoryNode, const SegmentedLevel, llfs::PageLoader> scanner{
-      node,
-      *this,
-      page_loader,
-      llfs::PinPageToJob::kDefault,
-      llfs::PageCacheOvercommit::not_allowed(),
-  };
-
-  Optional<std::string> prev_slice_max_key;
-  usize item_i = 0;
-
-  for (;;) {
-    Optional<EditSlice> edit_slice = scanner.next();
-
-    if (!edit_slice) {
-      break;
-    }
-
-    batt::case_of(*edit_slice, [&](const auto& slice_impl) {
-      if (slice_impl.empty()) {
-        return;
-      }
-      if (prev_slice_max_key) {
-        BATT_CHECK_LE(*prev_slice_max_key, get_key(slice_impl.front()));
-      }
-
-      prev_slice_max_key = std::string{get_key(slice_impl.back())};
-      item_i += slice_impl.size();
-    });
-  }
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-BoxedSeq<EditSlice> InMemoryNode::UpdateBuffer::SegmentedLevel::to_boxed_seq(
-    const InMemoryNode& node,
-    BatchUpdateContext& update_context,
-    Status& segment_load_status,
-    i32 min_pivot_i,
-    bool only_pivot,
-    Optional<KeyView> min_key) const
-{
-  if (only_pivot && !this->is_pivot_active(min_pivot_i)) {
-    return seq::Empty<EditSlice>{}  //
-           | seq::boxed();
-  }
-  return SegmentedLevelScanner<const InMemoryNode, const SegmentedLevel, llfs::PageLoader>{
-             node,
-             *this,
-             update_context.page_loader,
-             llfs::PinPageToJob::kDefault,
-             update_context.overcommit,
-             segment_load_status,
-             min_pivot_i,
-             min_key}  //
-         | seq::boxed();
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-bool InMemoryNode::UpdateBuffer::SegmentedLevel::set_pivot_items_flushed(
-    const InMemoryNode& node,
-    BatchUpdateContext& update_context,
-    usize pivot_i,
-    const CInterval<KeyView>& flush_key_crange,
-    Status& segment_load_status)
-{
-  segment_load_status.Update(
-      in_segmented_level(node, *this, update_context.page_loader, update_context.overcommit)
-          .drop_key_range(pivot_i, flush_key_crange.upper_bound));
-
-  return this->empty();
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-bool InMemoryNode::UpdateBuffer::SegmentedLevel::set_pivot_completely_flushed(usize pivot_i)
-{
-  for (usize segment_i = 0; segment_i < this->segment_count();) {
-    Segment& segment = this->get_segment(segment_i);
-
-    segment.set_pivot_active(pivot_i, false);
-
-    if (segment.is_inactive()) {
-      this->drop_segment(segment_i);
-    } else {
-      ++segment_i;
-    }
-  }
-
-  return this->empty();
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-usize InMemoryNode::UpdateBuffer::SegmentedLevel::segment_filter_cut_points() const
-{
-  usize n = 0;
-  for (const Segment& segment : this->segments) {
-    n += segment.num_cut_points();
-  }
-  return n;
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-void InMemoryNode::UpdateBuffer::SegmentedLevel::push_front_pivots(usize node_pivot_count)
-{
-  for (usize segment_i = 0; segment_i < this->segment_count(); ++segment_i) {
-    Segment& segment = this->segments[segment_i];
-    segment.active_pivots.push_front_pivots(node_pivot_count);
-  };
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-StatusOr<ValueView> InMemoryNode::UpdateBuffer::SegmentedLevel::find_key(const InMemoryNode& node,
-                                                                         KeyQuery& query,
-                                                                         i32 key_pivot_i) const
-{
-  return in_segmented_level(node, *this, *query.page_loader, query.overcommit())
-      .find_key(key_pivot_i, query);
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-Level InMemoryNode::UpdateBuffer::SegmentedLevel::merge(Level&& sibling_level,
-                                                        usize node_pivot_count) &&
-{
-  return batt::case_of(
-      sibling_level,
-      [&](EmptyLevel&) -> Level {
-        return *this;
-      },
-      [&](MergedLevel& right_merged_level) -> Level {
-        HybridLevel new_hybrid_level;
-        new_hybrid_level.add_new_sub_level(std::move(*this));
-        new_hybrid_level.add_new_sub_level(std::move(right_merged_level));
-
-        return new_hybrid_level;
-      },
-      [&](SegmentedLevel& right_segmented_level) -> Level {
-        // First shift the right level's bit set to the left by the number of pivots
-        // in the left node.
-        //
-        right_segmented_level.push_front_pivots(node_pivot_count);
-
-        this->segments.insert(this->segments.end(),
-                              std::make_move_iterator(right_segmented_level.segments.begin()),
-                              std::make_move_iterator(right_segmented_level.segments.end()));
-
-        // Next we need to find potential duplicate segments that have arisen from a previous
-        // split. For these duplicates, we need to merge their metadata.
-        //
-        std::unordered_map<llfs::PageId, usize, llfs::PageId::Hash> page_id_to_index;
-        SmallVec<bool, 32> should_remove(this->segments.size(), false);
-
-        for (usize i = 0; i < this->segments.size(); ++i) {
-          const llfs::PageId current_page_id = this->segments[i].page_id_slot.page_id;
-
-          auto iter = page_id_to_index.find(current_page_id);
-          if (iter != page_id_to_index.end()) {
-            // Found a duplicate, so we need to merge the segment metadata.
-            //
-            const usize first_index = iter->second;
-
-            // Merge active pivots using bitwise OR.
-            //
-            this->segments[first_index].active_pivots |= this->segments[i].active_pivots;
-
-            // Merge the filters.
-            //
-            this->segments[first_index].filter.merge(this->segments[i].filter);
-
-            // Mark this duplicate for removal.
-            //
-            should_remove[i] = true;
-          } else {
-            // We found the first occurrence of this page id, start tracking it.
-            //
-            page_id_to_index[current_page_id] = i;
-          }
-        }
-
-        // Remove duplicate segments, which have been marked for removal.
-        //
-        this->segments.erase(std::remove_if(this->segments.begin(),
-                                            this->segments.end(),
-                                            [&, i = usize{0}](const Segment&) mutable {
-                                              return should_remove[i++];
-                                            }),
-                             this->segments.end());
-
-        return *this;
-      },
-      [&](HybridLevel& right_hybrid_level) -> Level {
-        HybridLevel new_hybrid_level;
-        new_hybrid_level.add_new_sub_level(std::move(*this));
-
-        new_hybrid_level.add_new_sub_level(std::move(right_hybrid_level), node_pivot_count);
-
-        return new_hybrid_level;
-      });
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-void InMemoryNode::UpdateBuffer::HybridLevel::push_front_pivots(usize node_pivot_count)
-{
-  for (auto& sub_level : this->levels) {
-    if (batt::is_case<SegmentedLevel>(sub_level)) {
-      SegmentedLevel& segmented_sub_level = std::get<SegmentedLevel>(sub_level);
-      segmented_sub_level.push_front_pivots(node_pivot_count);
-    }
-  }
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-BoxedSeq<EditSlice> InMemoryNode::UpdateBuffer::HybridLevel::to_boxed_seq(
-    const InMemoryNode& node,
-    BatchUpdateContext& update_context,
-    Status& segment_load_status,
-    i32 min_pivot_i,
-    bool only_pivot,
-    Optional<KeyView> min_key) const
-{
-  return as_seq(this->levels) |
-         seq::map([&](const std::variant<MergedLevel, SegmentedLevel>& v) -> BoxedSeq<EditSlice> {
-           return batt::case_of(
-               v,
-               [&](const MergedLevel& merged_level) -> BoxedSeq<EditSlice> {
-                 return merged_level.to_boxed_seq(node, min_pivot_i);
-               },
-               [&](const SegmentedLevel& segmented_level) -> BoxedSeq<EditSlice> {
-                 return segmented_level.to_boxed_seq(node,
-                                                     update_context,
-                                                     segment_load_status,
-                                                     min_pivot_i,
-                                                     only_pivot,
-                                                     min_key);
-               });
-         }) |
-         seq::flatten() | seq::boxed();
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-bool InMemoryNode::UpdateBuffer::HybridLevel::set_pivot_items_flushed(
-    const InMemoryNode& node,
-    BatchUpdateContext& update_context,
-    usize pivot_i,
-    const CInterval<KeyView>& flush_key_crange,
-    Status segment_load_status)
-{
-  for (auto iter = this->levels.begin(); iter != this->levels.end();) {
-    bool sub_level_is_now_empty = false;
-
-    batt::case_of(
-        *iter,
-        [&](MergedLevel& merged_sub_level) {
-          sub_level_is_now_empty = merged_sub_level.set_items_flushed(flush_key_crange);
-        },
-        [&](SegmentedLevel& segmented_sub_level) {
-          sub_level_is_now_empty = segmented_sub_level.set_pivot_items_flushed(node,
-                                                                               update_context,
-                                                                               pivot_i,
-                                                                               flush_key_crange,
-                                                                               segment_load_status);
-        });
-
-    if (sub_level_is_now_empty) {
-      iter = this->levels.erase(iter);
-    } else {
-      ++iter;
-    }
-  }
-
-  return this->empty();
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-bool InMemoryNode::UpdateBuffer::HybridLevel::set_pivot_completely_flushed(
-    usize pivot_i,
-    const Interval<KeyView>& pivot_key_range)
-{
-  for (auto iter = this->levels.begin(); iter != this->levels.end();) {
-    bool sub_level_is_now_empty = false;
-
-    batt::case_of(
-        *iter,
-        [&](MergedLevel& merged_sub_level) {
-          sub_level_is_now_empty = merged_sub_level.set_items_flushed(pivot_key_range);
-        },
-        [&](SegmentedLevel& segmented_sub_level) {
-          sub_level_is_now_empty = segmented_sub_level.set_pivot_completely_flushed(pivot_i);
-        });
-
-    if (sub_level_is_now_empty) {
-      iter = this->levels.erase(iter);
-    } else {
-      ++iter;
-    }
-  }
-
-  return this->empty();
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-usize InMemoryNode::UpdateBuffer::HybridLevel::segment_count(const TreeOptions& tree_options) const
-{
-  usize total = 0;
-
-  for (auto iter = this->levels.begin(); iter != this->levels.end(); ++iter) {
-    total += batt::case_of(
-        *iter,
-        [&](const MergedLevel& merged_sub_level) -> usize {
-          return merged_sub_level.estimate_segment_count(tree_options);
-        },
-        [&](const SegmentedLevel& segmented_sub_level) -> usize {
-          return segmented_sub_level.segment_count();
-        });
-  }
-
-  return total;
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-void InMemoryNode::UpdateBuffer::HybridLevel::drop_before_pivot(i32 pivot_i,
-                                                                const KeyView& pivot_key,
-                                                                llfs::PageLoader& page_loader,
-                                                                const TreeOptions& tree_options)
-{
-  for (auto iter = this->levels.begin(); iter != this->levels.end(); ++iter) {
-    batt::case_of(*iter, [&](auto& sub_level) {
-      sub_level.drop_before_pivot(pivot_i, pivot_key, page_loader, tree_options);
-    });
-  }
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-void InMemoryNode::UpdateBuffer::HybridLevel::drop_after_pivot(i32 pivot_i,
-                                                               const KeyView& pivot_key,
-                                                               llfs::PageLoader& page_loader,
-                                                               const TreeOptions& tree_options)
-{
-  for (auto iter = this->levels.begin(); iter != this->levels.end(); ++iter) {
-    batt::case_of(*iter, [&](auto& sub_level) {
-      sub_level.drop_after_pivot(pivot_i, pivot_key, page_loader, tree_options);
-    });
-  }
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-Status InMemoryNode::UpdateBuffer::HybridLevel::split_pivot(
-    InMemoryNode& node,
-    BatchUpdateContext& update_context,
-    i32 pivot_i,
-    const Interval<KeyView>& pivot_key_range,
-    const KeyView& sibling_min_key)
-{
-  for (auto& sub_level : this->levels) {
-    if (batt::is_case<SegmentedLevel>(sub_level)) {
-      SegmentedLevel& segmented_sub_level = std::get<SegmentedLevel>(sub_level);
-      BATT_REQUIRE_OK(in_segmented_level(node,
-                                         segmented_sub_level,
-                                         update_context.page_loader,
-                                         update_context.overcommit)  //
-                          .split_pivot(pivot_i, pivot_key_range, sibling_min_key));
-    }
-  }
-
-  return OkStatus();
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-void InMemoryNode::UpdateBuffer::HybridLevel::merge_pivots(InMemoryNode& node,
-                                                           BatchUpdateContext& update_context,
-                                                           i32 left_pivot_i,
-                                                           i32 right_pivot_i)
-{
-  for (auto& sub_level : this->levels) {
-    if (batt::is_case<SegmentedLevel>(sub_level)) {
-      SegmentedLevel& segmented_sub_level = std::get<SegmentedLevel>(sub_level);
-      in_segmented_level(*this,
-                         segmented_sub_level,
-                         update_context.page_loader,
-                         update_context.overcommit)
-          .merge_pivots(left_pivot_i, right_pivot_i);
-    }
-  }
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-StatusOr<ValueView> InMemoryNode::UpdateBuffer::HybridLevel::find_key(const InMemoryNode& node,
-                                                                      KeyQuery& query,
-                                                                      i32 key_pivot_i) const
-{
-  StatusOr<ValueView> result{Status{batt::StatusCode::kNotFound}};
-
-  // TODO [vsilai 2026-04-28]: Implement a way to avoid iterating through all sub-levels.
-  //
-  for (const auto& sub_level : this->levels) {
-    result = batt::case_of(
-        sub_level,
-        [&](const MergedLevel& merged_sub_level) -> StatusOr<ValueView> {
-          return merged_sub_level.find_key(query.key());
-        },
-        [&](const SegmentedLevel& segmented_sub_level) -> StatusOr<ValueView> {
-          return segmented_sub_level.find_key(node, query, key_pivot_i);
-        });
-
-    if (result.ok() || result.status() != batt::StatusCode::kNotFound) {
-      return result;
-    }
-  }
-
-  return result;
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-Level InMemoryNode::UpdateBuffer::HybridLevel::merge(Level&& sibling_level,
-                                                     usize node_pivot_count) &&
-{
-  batt::case_of(
-      sibling_level,
-      [&](EmptyLevel&) {
-      },
-      [&](MergedLevel& right_merged_level) {
-        this->add_new_sub_level(std::move(right_merged_level));
-      },
-      [&](SegmentedLevel& right_segmented_level) {
-        this->add_new_sub_level(std::move(right_segmented_level), node_pivot_count);
-      },
-      [&](HybridLevel& right_hybrid_level) {
-        this->add_new_sub_level(std::move(right_hybrid_level), node_pivot_count);
-      });
-
-  return *this;
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-StatusOr<usize> InMemoryNode::UpdateBuffer::HybridLevel::start_serialize(
-    const InMemoryNode& node,
-    TreeSerializeContext& context)
-{
-  usize total_segment_count = 0;
-
-  for (auto& sub_level : this->levels) {
-    StatusOr<usize> sub_level_segment_count = batt::case_of(
-        sub_level,
-        [&](MergedLevel& merged_sub_level) -> StatusOr<usize> {
-          BATT_ASSIGN_OK_RESULT(usize sub_level_total,
-                                merged_sub_level.start_serialize(node, context));
-          return sub_level_total;
-        },
-        [&](SegmentedLevel& segmented_sub_level) -> StatusOr<usize> {
-          return segmented_sub_level.segment_count();
-        });
-
-    BATT_REQUIRE_OK(sub_level_segment_count);
-
-    total_segment_count += *sub_level_segment_count;
-  }
-
-  return total_segment_count;
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-StatusOr<SegmentedLevel> InMemoryNode::UpdateBuffer::HybridLevel::finish_serialize(
-    const InMemoryNode& node,
-    TreeSerializeContext& context)
-{
-  static Metrics& r_metrics = InMemoryNode::metrics();
-
-  SegmentedLevel final_segmented_level;
-
-  for (auto& sub_level : this->levels) {
-    r_metrics.serialized_nonempty_level_count.add(1);
-
-    StatusOr<SegmentedLevel> new_segmented_sub_level = batt::case_of(
-        sub_level,
-        [&](MergedLevel& merged_sub_level) -> StatusOr<SegmentedLevel> {
-          return merged_sub_level.finish_serialize(node, context);
-        },
-        [&](SegmentedLevel& segmented_sub_level) -> StatusOr<SegmentedLevel> {
-          r_metrics.serialized_buffer_segment_count.add(segmented_sub_level.segment_count());
-
-          return {std::move(segmented_sub_level)};
-        });
-
-    BATT_REQUIRE_OK(new_segmented_sub_level);
-
-    final_segmented_level.segments.insert(
-        final_segmented_level.segments.end(),
-        std::make_move_iterator(new_segmented_sub_level->segments.begin()),
-        std::make_move_iterator(new_segmented_sub_level->segments.end()));
-  }
-
-  return {std::move(final_segmented_level)};
-}
-
-//=#=#==#==#===============+=+=+=+=++=++++++++++++++-++-+--+-+----+---------------
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-void InMemoryNode::UpdateBuffer::Segment::check_invariants(const char* file, int line) const
-{
-  BATT_CHECK(this->filter.check_invariants()) << BATT_INSPECT(file) << BATT_INSPECT(line);
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-void InMemoryNode::UpdateBuffer::Segment::insert_pivot(i32 pivot_i, bool is_active)
-{
-  this->check_invariants(__FILE__, __LINE__);
-  auto on_scope_exit = batt::finally([&] {
-    this->check_invariants(__FILE__, __LINE__);
-  });
-
-  this->active_pivots.insert(pivot_i, is_active);
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-void InMemoryNode::UpdateBuffer::Segment::remove_pivot(i32 pivot_i)
-{
-  this->check_invariants(__FILE__, __LINE__);
-  auto on_scope_exit = batt::finally([&] {
-    this->check_invariants(__FILE__, __LINE__);
-  });
-
-  this->active_pivots.remove(pivot_i);
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-void InMemoryNode::UpdateBuffer::Segment::pop_front_pivots(i32 count)
-{
-  BATT_CHECK_LT(count, 64);
-
-  // Shift the active pivot sets down by count.
-  //
-  this->active_pivots.pop_front_pivots(count);
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-bool InMemoryNode::UpdateBuffer::Segment::is_inactive() const
-{
-  const bool inactive = this->active_pivots.is_empty();
-  if (inactive) {
-    Slice<const Interval<u32>> live_ranges = this->filter.live();
-    BATT_CHECK_EQ(live_ranges.size(), 1) << BATT_INSPECT(live_ranges);
-    BATT_CHECK_EQ(live_ranges[0].upper_bound, PiecewiseFilter<u32>::kMaxUpperBound)
-        << BATT_INSPECT(live_ranges);
-  }
-  return inactive;
-}
-
-//=#=#==#==#===============+=+=+=+=++=++++++++++++++-++-+--+-+----+---------------
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
 SmallFn<void(std::ostream&)> InMemoryNode::UpdateBuffer::dump() const
 {
   return [this](std::ostream& out) {
@@ -2472,93 +1679,6 @@ SmallFn<void(std::ostream&)> InMemoryNode::UpdateBuffer::dump() const
       });
     }
     out << "},}";
-  };
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-Level InMemoryNode::UpdateBuffer::EmptyLevel::merge(Level&& sibling_level,
-                                                    usize node_pivot_count) &&
-{
-  return batt::case_of(
-      sibling_level,
-      [&](EmptyLevel&) -> Level {
-        return *this;
-      },
-      [&](MergedLevel& right_merged_level) -> Level {
-        return std::move(right_merged_level);
-      },
-      [&](SegmentedLevel& right_segmented_level) -> Level {
-        right_segmented_level.push_front_pivots(node_pivot_count);
-
-        return std::move(right_segmented_level);
-      },
-      [&](HybridLevel& right_hybrid_level) -> Level {
-        right_hybrid_level.push_front_pivots(node_pivot_count);
-        return std::move(right_hybrid_level);
-      });
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-SmallFn<void(std::ostream&)> InMemoryNode::UpdateBuffer::EmptyLevel::dump() const
-{
-  return [](std::ostream& out) {
-    out << "EmptyLevel{}";
-  };
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-SmallFn<void(std::ostream&)> InMemoryNode::UpdateBuffer::MergedLevel::dump() const
-{
-  return [this](std::ostream& out) {
-    out << "MergedLevel{" << this->result_set.debug_dump("    ") << "\n}";
-  };
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-SmallFn<void(std::ostream&)> InMemoryNode::UpdateBuffer::SegmentedLevel::dump() const
-{
-  return [this](std::ostream& out) {
-    out << "SegmentedLevel{\n";
-    for (const Segment& segment : this->segments) {
-      out << "    " << segment.dump(/*multi_line=*/false) << ",\n";
-    }
-    out << "  }";
-  };
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-SmallFn<void(std::ostream&)> InMemoryNode::UpdateBuffer::HybridLevel::dump() const
-{
-  return [this](std::ostream& out) {
-    out << "HybridLevel{\n";
-    for (auto iter = this->levels.begin(); iter != this->levels.end(); ++iter) {
-      batt::case_of(*iter, [](const auto& sub_level) {
-        sub_level.dump();
-      });
-    }
-    out << "  }";
-  };
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-SmallFn<void(std::ostream&)> InMemoryNode::UpdateBuffer::Segment::dump(bool multi_line) const
-{
-  return [this, multi_line](std::ostream& out) {
-    if (multi_line) {
-      out << "Segment:" << std::endl
-          << "   active=" << this->active_pivots.printable() << std::endl
-          << "   filter=" << this->filter.dump() << std::endl
-          << std::endl;
-    } else {
-      out << "Segment{.active=" << this->active_pivots.printable()
-          << ", .filter=" << this->filter.dump() << ",}";
-    }
   };
 }
