+++ src/turtle_kv/tree/update_buffer_levels.hpp
@@ -1,0 +1,17 @@
+#pragma once
+
+#include <variant>
+
+namespace turtle_kv {
+
+struct InMemoryNodeEmptyLevel;
+struct InMemoryNodeMergedLevel;
+struct InMemoryNodeSegmentedLevel;
+struct InMemoryNodeHybridLevel;
+
+using InMemoryNodeLevel = std::variant<InMemoryNodeEmptyLevel,
+                                       InMemoryNodeMergedLevel,
+                                       InMemoryNodeSegmentedLevel,
+                                       InMemoryNodeHybridLevel>;
+
+}  // namespace turtle_kv
